fix: remove duplicate /users/:userId/achievements page
UserProfilePage already rendered the full badge catalog with earned
status; the dedicated achievements page, its route, and the
GetUserAchievements endpoint showed identical content and only added
an extra fetch and dead-end "View all achievements" link.

Closes #1032
