+++ backend/src/Api/wwwroot/openapi-v1.json
@@ -5996,50 +5996,6 @@
         }
       }
     },
-    "/v1/users/{userId}/achievements": {
-      "get": {
-        "tags": [
-          "Achievements"
-        ],
-        "operationId": "GetUserAchievements",
-        "parameters": [
-          {
-            "name": "userId",
-            "in": "path",
-            "required": true,
-            "schema": {
-              "type": "string",
-              "format": "uuid"
-            }
-          }
-        ],
-        "responses": {
-          "200": {
-            "description": "OK",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "type": "array",
-                  "items": {
-                    "$ref": "#/components/schemas/AchievementSummary"
-                  }
-                }
-              }
-            }
-          },
-          "500": {
-            "description": "Internal Server Error",
-            "content": {
-              "application/problem+json": {
-                "schema": {
-                  "$ref": "#/components/schemas/ProblemDetails"
-                }
-              }
-            }
-          }
-        }
-      }
-    },
     "/v1/me/achievements": {
       "get": {
         "tags": [
