+++ frontend/src/client/api-client.ts
@@ -5341,52 +5341,6 @@ export class EinsatzbereitApi {
     /**
      * @return OK
      */
-    getUserAchievements(userId: string, signal?: AbortSignal): Promise<AchievementSummary[]> {
-        let url_ = this.baseUrl + "/v1/users/{userId}/achievements";
-        if (userId === undefined || userId === null)
-            throw new globalThis.Error("The parameter 'userId' must be defined.");
-        url_ = url_.replace("{userId}", encodeURIComponent("" + userId));
-        url_ = url_.replace(/[?&]$/, "");
-
-        let options_: RequestInit = {
-            method: "GET",
-            signal,
-            headers: {
-                "Accept": "application/json"
-            }
-        };
-
-        return this.http.fetch(url_, options_).then((_response: Response) => {
-            return this.processGetUserAchievements(_response);
-        });
-    }
-
-    protected processGetUserAchievements(response: Response): Promise<AchievementSummary[]> {
-        const status = response.status;
-        let _headers: any = {}; if (response.headers && response.headers.forEach) { response.headers.forEach((v: any, k: any) => _headers[k] = v); };
-        if (status === 200) {
-            return response.text().then((_responseText) => {
-            let result200: any = null;
-            result200 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver) as AchievementSummary[];
-            return result200;
-            });
-        } else if (status === 500) {
-            return response.text().then((_responseText) => {
-            let result500: any = null;
-            result500 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver) as ProblemDetails;
-            return throwException("Internal Server Error", status, _responseText, _headers, result500);
-            });
-        } else if (status !== 200 && status !== 204) {
-            return response.text().then((_responseText) => {
-            return throwException("An unexpected server error occurred.", status, _responseText, _headers);
-            });
-        }
-        return Promise.resolve<AchievementSummary[]>(null as any);
-    }
-
-    /**
-     * @return OK
-     */
     getMyAchievements(signal?: AbortSignal): Promise<AchievementSummary[]> {
         let url_ = this.baseUrl + "/v1/me/achievements";
         url_ = url_.replace(/[?&]$/, "");
